Fix: Reducing default number of decimal places from 24 to 17 for native formatted floats.
The current MDN documentation says that max value is 100. But in Selenium testing in Webchart, I was getting this error:

    RangeError: invalid digits value: 24

I also found reports online that the max value used to be 20. Unfortunately MDN doesn't always do a good job of tracking changes in browser implementations over time, so I can't confirm that. But in any case, 17 digits is enough to represent all IEEE-754 64-bit floats.