Revert "Fix: Pinning a column now disables the 'limit' feature, working around a Tabletool bug."
This reverts commit 5cb4ef1712404124138918f946231e8e0497c80a.

It turned out that this workaround was too onerous for customers particularly when they used DataVis
mainly to generate a CSV file of the data, which would previously work without showing all the
data. For example, they could create a CSV file with 50,000 rows whereas actually rendering all
those rows would cause the browser to run out of memory and crash. We will try another way.